Move favorites tests into sub directory
