+++ tests/favorites/view-favorite-by-id.spec.js
@@ -1,18 +1,18 @@
 var shell = require('shelljs');
 var request = require("supertest");
-var app = require('../app');
+var app = require('../../app');
 
 const environment = process.env.NODE_ENV || 'test';
-const configuration = require('../knexfile')[environment];
+const configuration = require('../../knexfile')[environment];
 const database = require('knex')(configuration);
 
 describe('Test the favorites path', () => {
   beforeEach(async () => {
-    await database.raw('truncate table favorites cascade');
+    await database.raw('TRUNCATE TABLE favorites RESTART IDENTITY CASCADE');
   });
 
-  afterEach(() => {
-    database.raw('truncate table favorites cascade');
+  afterEach(async () => {
+    await database.raw('TRUNCATE TABLE favorites RESTART IDENTITY CASCADE');
   });
 
   test('It should respond to a GET request for a favorite :id with that favorite' , async () => {
